Extract `DefaultDropdownMenu` to the `PageComponents.kt`.
